requirements.txt
requests
customtkinter
Pillow
